# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Fullstack project tracker", "version": "0.1.0"}, "paths": {"/api/v1/projects/": {"get": {"tags": ["projects"], "summary": "Read Projects", "description": "Read projects.", "operationId": "read_projects_api_v1_projects__get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectListRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["projects"], "summary": "Create Project", "description": "Create project.", "operationId": "create_project_api_v1_projects__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{id}": {"get": {"tags": ["projects"], "summary": "Read Project By Id", "description": "Read a specific project by id.", "operationId": "read_project_by_id_api_v1_projects__id__get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["projects"], "summary": "Update Project", "description": "Update a project.", "operationId": "update_project_api_v1_projects__id__patch", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["projects"], "summary": "Delete Project", "description": "Delete a project", "operationId": "delete_project_api_v1_projects__id__delete", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/logs": {"get": {"tags": ["project-change-logs"], "summary": "Read Project Logs", "description": "Read project logs.", "operationId": "read_project_logs_api_v1_projects__project_id__logs_get", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Project Id"}}, {"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectChangeLogListRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/comments": {"get": {"tags": ["project-comments"], "summary": "Read Project Comments", "description": "Read project comments.", "operationId": "read_project_comments_api_v1_projects__project_id__comments_get", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Project Id"}}, {"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCommentListRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["project-comments"], "summary": "Create Project Comment", "operationId": "create_project_comment_api_v1_projects__project_id__comments_post", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCommentCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCommentRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/comments/{id}": {"delete": {"tags": ["project-comments"], "summary": "Delete Project", "description": "Delete a project", "operationId": "delete_project_api_v1_projects__project_id__comments__id__delete", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Id"}}, {"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Project Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ProjectChangeLog": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "project_id": {"type": "string", "format": "uuid", "title": "Project Id"}, "changed_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Changed At"}, "field_name": {"type": "string", "title": "Field Name"}, "old_value": {"type": "string", "title": "Old Value"}, "new_value": {"type": "string", "title": "New Value"}}, "type": "object", "required": ["project_id", "field_name", "old_value", "new_value"], "title": "ProjectChangeLog"}, "ProjectChangeLogListRead": {"properties": {"data": {"items": {"$ref": "#/components/schemas/ProjectChangeLog"}, "type": "array", "title": "Data"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["data", "count"], "title": "ProjectChangeLogListRead"}, "ProjectCommentCreate": {"properties": {"content": {"type": "string", "maxLength": 255, "title": "Content"}}, "type": "object", "required": ["content"], "title": "ProjectCommentCreate"}, "ProjectCommentListRead": {"properties": {"data": {"items": {"$ref": "#/components/schemas/ProjectCommentRead"}, "type": "array", "title": "Data"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["data", "count"], "title": "ProjectCommentListRead"}, "ProjectCommentRead": {"properties": {"content": {"type": "string", "maxLength": 255, "title": "Content"}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "project_id": {"type": "string", "format": "uuid", "title": "Project Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["content", "id", "project_id", "created_at"], "title": "ProjectCommentRead"}, "ProjectCreate": {"properties": {"name": {"type": "string", "maxLength": 255, "title": "Name"}, "description": {"type": "string", "maxLength": 255, "title": "Description"}, "status": {"$ref": "#/components/schemas/ProjectStatus"}}, "type": "object", "required": ["name", "description", "status"], "title": "ProjectCreate"}, "ProjectListRead": {"properties": {"data": {"items": {"$ref": "#/components/schemas/ProjectRead"}, "type": "array", "title": "Data"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["data", "count"], "title": "ProjectListRead"}, "ProjectRead": {"properties": {"name": {"type": "string", "maxLength": 255, "title": "Name"}, "description": {"type": "string", "maxLength": 255, "title": "Description"}, "status": {"$ref": "#/components/schemas/ProjectStatus"}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["name", "description", "status", "id", "created_at", "updated_at"], "title": "ProjectRead"}, "ProjectStatus": {"type": "string", "enum": ["planned", "active", "completed"], "title": "ProjectStatus"}, "ProjectUpdate": {"properties": {"name": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Name"}, "description": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Description"}, "status": {"anyOf": [{"$ref": "#/components/schemas/ProjectStatus"}, {"type": "null"}]}}, "type": "object", "title": "ProjectUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}